AkiKyo/requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
arviz==0.12.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
cftime==1.6.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
DateTime==4.3
deprecat==2.1.1
dill==0.3.4
discord==1.7.3
discord.py==1.7.3
docopt==0.6.2
fastprogress==1.0.2
filelock==3.7.0
Flask==2.1.2
fonttools==4.33.3
future==0.18.2
geographiclib==1.52
geopy==2.2.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
line-bot-sdk==2.2.1
MarkupSafe==2.1.1
matplotlib==3.5.2
multidict==5.2.0
netCDF4==1.5.8
numpy==1.22.1
outcome==1.1.0
packaging==21.3
pandas==1.4.2
patsy==0.5.2
pefile==2021.9.3
pendulum==2.1.2
Pillow==9.1.0
pipreqs==0.4.11
pycparser==2.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
pymc3==3.11.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pytzdata==2020.1
pywin32-ctypes==0.2.0
requests==2.27.1
requests-file==1.5.1
scipy==1.7.3
selenium==4.1.3
semver==2.13.0
Shapely==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Theano-PyMC==1.1.2
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.0.0
tzwhere==3.0.3
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
xarray==2022.3.0
xarray-einstats==0.2.2
yarg==0.1.9
yarl==1.7.2
zipp==3.8.0
zope.interface==5.4.0
